remove row_serde trait
