+++ src/storage/src/row_serde/mod.rs
@@ -15,47 +15,10 @@
 use std::collections::HashMap;
 use std::sync::Arc;
 
-use risingwave_common::array::Row;
 use risingwave_common::catalog::{ColumnDesc, ColumnId};
-use risingwave_common::error::Result;
-use risingwave_common::types::{DataType, VirtualNode};
-use risingwave_common::util::ordered::OrderedRowSerializer;
+use risingwave_common::types::DataType;
 
-use self::row_based_deserializer::RowBasedDeserializer;
-use self::row_based_serializer::RowBasedSerializer;
-
-pub mod row_based_deserializer;
-pub mod row_based_serializer;
 pub mod row_serde_util;
-
-#[derive(Clone)]
-pub struct RowBasedSerde;
-
-impl RowSerde for RowBasedSerde {
-    type Deserializer = RowBasedDeserializer;
-    type Serializer = RowBasedSerializer;
-}
-
-pub type KeyBytes = Vec<u8>;
-pub type ValueBytes = Vec<u8>;
-
-/// `RowSerialize` defines an interface for encoding a key row into kv storage.
-pub trait RowSerialize: Clone {
-    /// Constructs a new serializer.
-    fn create_row_serializer(
-        pk_indices: &[usize],
-        column_descs: &[ColumnDesc],
-        column_ids: &[ColumnId],
-    ) -> Self;
-
-    /// Serialize key and value.
-    fn serialize(
-        &mut self,
-        vnode: VirtualNode,
-        pk: &[u8],
-        row: Row,
-    ) -> Result<(KeyBytes, ValueBytes)>;
-}
 
 /// `ColumnDescMapping` is the record mapping from [`ColumnDesc`], [`ColumnId`], and is used in both
 /// cell-based encoding and row-based encoding.
@@ -74,42 +37,64 @@ pub struct ColumnDescMapping {
     pub output_index: Vec<usize>,
 }
 
-/// `RowDeserialize` defines an interface for decoding a key row from kv storage.
-pub trait RowDeserialize {
-    /// Constructs a new deserializer.
-    fn create_row_deserializer(column_mapping: Arc<ColumnDescMapping>) -> Self;
-
-    /// When we encounter a new key, we can be sure that the previous row has been fully
-    /// deserialized. Then we return the key and the value of the previous row.
-    fn deserialize(
-        &mut self,
-        raw_key: impl AsRef<[u8]>,
-        value: impl AsRef<[u8]>,
-    ) -> Result<(VirtualNode, Vec<u8>, Row)>;
-}
-
-/// `RowSerde` provides the ability to convert between Row and KV entry.
-pub trait RowSerde: Send + Sync + Clone {
-    type Serializer: RowSerialize + Send;
-    type Deserializer: RowDeserialize + Send;
-
-    /// `create_serializer` will create a row serializer to convert row into KV pairs.
-    fn create_serializer(
-        pk_indices: &[usize],
-        column_descs: &[ColumnDesc],
-        column_ids: &[ColumnId],
-    ) -> Self::Serializer {
-        RowSerialize::create_row_serializer(pk_indices, column_descs, column_ids)
+#[allow(clippy::len_without_is_empty)]
+impl ColumnDescMapping {
+    fn new_inner(
+        output_columns: Vec<ColumnDesc>,
+        all_data_types: Vec<DataType>,
+        output_index: Vec<usize>,
+    ) -> Arc<Self> {
+        let output_id_to_index = output_columns
+            .iter()
+            .enumerate()
+            .map(|(index, d)| (d.column_id, index))
+            .collect();
+        Self {
+            output_columns,
+            output_id_to_index,
+            all_data_types,
+            output_index,
+        }
+        .into()
     }
 
-    /// `create_deserializer` will create a row deserializer to convert KV pairs into row.
-    fn create_deserializer(column_mapping: Arc<ColumnDescMapping>) -> Self::Deserializer {
-        RowDeserialize::create_row_deserializer(column_mapping)
+    /// Create a mapping with given `output_columns`.
+    pub fn new(output_columns: Vec<ColumnDesc>) -> Arc<Self> {
+        let all_data_types = output_columns.iter().map(|d| d.data_type.clone()).collect();
+        let output_index: Vec<usize> = output_columns
+            .iter()
+            .map(|c| c.column_id.get_id() as usize)
+            .collect();
+        Self::new_inner(output_columns, all_data_types, output_index)
     }
-}
 
-pub fn serialize_pk(pk: &Row, serializer: &OrderedRowSerializer) -> Vec<u8> {
-    let mut result = vec![];
-    serializer.serialize(pk, &mut result);
-    result
+    /// Create a mapping with given `table_columns` projected on the `column_ids`.
+    pub fn new_partial(table_columns: &[ColumnDesc], output_column_ids: &[ColumnId]) -> Arc<Self> {
+        let all_data_types = table_columns.iter().map(|d| d.data_type.clone()).collect();
+        let mut table_columns = table_columns
+            .iter()
+            .enumerate()
+            .map(|(index, c)| (c.column_id, (c.clone(), index)))
+            .collect::<HashMap<_, _>>();
+        let (output_columns, output_index): (
+            Vec<risingwave_common::catalog::ColumnDesc>,
+            Vec<usize>,
+        ) = output_column_ids
+            .iter()
+            .map(|id| table_columns.remove(id).unwrap())
+            .unzip();
+        Self::new_inner(output_columns, all_data_types, output_index)
+    }
+
+    /// Get the [`ColumnDesc`] and its index in the output with given `id`.
+    pub fn get(&self, id: ColumnId) -> Option<(&ColumnDesc, usize)> {
+        self.output_id_to_index
+            .get(&id)
+            .map(|&index| (&self.output_columns[index], index))
+    }
+
+    /// Get the length of output columns.
+    pub fn len(&self) -> usize {
+        self.output_columns.len()
+    }
 }
